FSPT-189: Set test client `set_cookie` `origin_only` to False
Our Authenticator and Assess Test Clients implement their own version
of the set_cookie method so that we don't have to explicitly pass the
domain every time we want to set the cookie in a test or request.

However the default implementation of this method (see
https://werkzeug.palletsprojects.com/en/stable/test/#werkzeug.test.Client.set_cookie)
sets `origin_only` to True, meaning the domain must be an exact match to
the request in order for a cookie to be set on the test request. This
doesn't work with our implementation as we want a cookie to be set
across the services.

This commit switches our implementation of `set_cookie` to default
`origin_only` to False.