FSPT-189: Update authenticator template paths
- Consolidate all authenticator templates into frontend/authenticator
folder so that host matching template rendering works
- Update all template and render paths
- Update error route handling to include service desk URL
- app.py uses the `inject_content_urls` method as a context processor which
calls `find_fund_and_round_in_request` on apply - the issue when the host
is authenticator is that there is no rund and round in the view_args and
these fund and round checks assume that request.view_args is not none and
calls a .get() on the view_args which threw a 500. This commit reworks
these functions to account for view_args being empty.